pyproject.toml
[tool.poetry]
name = "morgandev"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
discord = "^1.0.1"
python-dotenv = "^0.15.0"
d20 = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
